fix: resolve build errors from Double→Long refactor
- DialogComponents: onConfirm callback type Double → Long
- DashboardComponents: add .toDouble() for Long/Long division ratios
- Fix 0.0 → 0L literals in budget comparisons

+++ app/src/main/java/com/example/ui/common/DialogComponents.kt
@@ -417,7 +417,7 @@ fun MonthlyBudgetDialog(
     currentBudgetLimit: Long,
     categoryBudgets: Map<String, Long>,
     expenseCategories: List<String>,
-    onConfirm: (Double, Map<String, Double>) -> Unit,
+    onConfirm: (Long, Map<String, Long>) -> Unit,
     onDismiss: () -> Unit
 ) {
     var budgetInput by remember {
